+++ ios/Classes/handlers/Handler.swift
@@ -112,6 +112,9 @@ enum AppHudMethod {
         case showPendingScreen
         case pendingScreenController
         case pendingRule
+        case checkRules
+        case submitPushNotificationsToken
+        case handlePushNotification
     }
 
     enum Attribution: String, CaseIterable {